+++ settingsdialog.h
@@ -65,6 +65,7 @@ private slots:
     void on_showDerivativeF0_stateChanged(int arg1);
     void on_markoutType_currentIndexChanged(int index);
     void on_checkBox_8_stateChanged(int arg1);
+    void on_autoMarking_stateChanged(int arg1);
 };
 
 #endif // SETTINGSDIALOG_H